Extracted common functionality into a trait.
Extracted common functionality into a trait used in both the tags class and the modifier.
